controller e rota para exclusão de um inquilino

+++ src/modules/tenents/services/DeleteTenantService.ts
@@ -3,7 +3,6 @@ import { inject, injectable } from 'tsyringe';
 import IUserRepository from '@modules/users/repositories/IUserRepository';
 import AppError from '@shared/errors/AppError';
 import ITenentRepository from '../repositories/ITenentRepository';
-import Tenent from '../infra/typeorm/entities/Tenents';
 
 interface IRequest {
   tenant_id: string;
@@ -20,16 +19,22 @@ class DeleteTenantService {
     private userRepository: IUserRepository,
   ) {}
 
-  public async execute({ tenant_id, user_id }: IRequest): Promise<Tenent> {
+  public async execute({ tenant_id, user_id }: IRequest): Promise<void> {
     const checkUser = await this.userRepository.findById(user_id);
 
     if (!checkUser) {
       throw new AppError('User not found');
     }
 
-    const tenant = await this.tenentRepository.delete(tenant_id, user_id);
+    const id = tenant_id;
 
-    return tenant;
+    const checkTenant = await this.tenentRepository.findById(id, user_id);
+
+    if (!checkTenant) {
+      throw new AppError('Tenant not found for authenticated user');
+    }
+
+    await this.tenentRepository.delete(tenant_id, user_id);
   }
 }
